added str to bsl_loader and removed loader.py
